fix: fix mentor comments
