+++ src/utils/common.js
@@ -1,4 +1,3 @@
-import { FILTERS } from '../const';
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
 const getRandomInteger = (a = 0, b = 1) => {
@@ -32,10 +31,5 @@ const calculateDuration = (minutes) => {
 
 const isPressedEscapeKey = (evt) => evt.key === 'Escape';
 
-const Filter = {
-  [FILTERS.WATCHLIST]: (films) => films.filter(({userDetails}) => userDetails.watchlist),
-  [FILTERS.HISTORY]: (films) => films.filter(({userDetails}) => userDetails.alreadyWatched),
-  [FILTERS.FAVORITES]: (films) => films.filter(({userDetails}) => userDetails.favorite)
-};
 
-export {getRandomInteger, generateRandomNameFromArray, getRandomDecimalNumber, calculateDuration, generateRandomArrayFromArray, isPressedEscapeKey, Filter};
+export {getRandomInteger, generateRandomNameFromArray, getRandomDecimalNumber, calculateDuration, generateRandomArrayFromArray, isPressedEscapeKey};
